+++ lib/minimizer/linmin.h
@@ -2,6 +2,6 @@
 #ifndef LINMIN_H
 #define LINMIN_H
 
-void linmin(GraphPointer gp, float (*e_fun)(GraphPointer), float *fret, VectorPointer gradient);
+void linmin(GraphPointer gp, double (*e_fun)(GraphPointer), double *fret, VectorPointer gradient);
 
 #endif
