+++ lib/minimizer/vector.h
@@ -6,13 +6,13 @@ typedef struct vector Vector, *VectorPointer;
 
 struct vector
 {
-    float x, y, len;
+    double x, y, len;
     int given_coords;
 };
 
-Vector Vector_initialize(const float x, const float y);
+Vector Vector_initialize(const double x, const double y);
 
-VectorPointer Vector_create(const float x, const float y);
+VectorPointer Vector_create(const double x, const double y);
 
 Vector Vector_zero();
 
@@ -24,17 +24,17 @@ Vector Vector_mult(Vector vec1, Vector vec2);
 
 Vector Vector_negate(Vector vec);
 
-Vector Vector_scalar_mult(Vector vec, float c);
+Vector Vector_scalar_mult(Vector vec, double c);
 
-Vector Vector_scalar_add(Vector vec, float c);
+Vector Vector_scalar_add(Vector vec, double c);
 
-float Vector_cross(Vector vec1, Vector vec2);
+double Vector_cross(Vector vec1, Vector vec2);
 
-float Vector_dot(Vector v1, Vector v2);
+double Vector_dot(Vector v1, Vector v2);
 
-float Vector_norm(Vector v);
+double Vector_norm(Vector v);
 
-float Vector_angle(Vector v1, Vector v2);
+double Vector_angle(Vector v1, Vector v2);
 
 int Vector_parallel(Vector v1, Vector v2);
