+++ lib/minimizer/vertex.h
@@ -24,8 +24,8 @@ VertexPointer Vertex_create(
         const int id, 
         const Vector pos, 
         const Vector grad, const Vector g, const Vector h,
-        const float wdth, const
-        float hght, 
+        const double wdth, 
+        const double hght, 
         const char type, 
         const int nv 
 );
@@ -35,7 +35,7 @@ int Vertex_zone_idx(const VertexPointer v);
 
 void Vertex_move(const VertexPointer vp, const Vector s);
 
-float Vertex_potential_energy(const VertexPointer vp);
+double Vertex_potential_energy(const VertexPointer vp);
 
 Vector Vertex_potential_gradient(const VertexPointer vp);
