+++ lib/minimizer/vertex_pair.h
@@ -2,7 +2,7 @@
 #ifndef VERTEX_PAIR_H
 #define VERTEX_PAIR_H
 
-float VertexPair_repulsion_energy(const Pair pr);
+double VertexPair_repulsion_energy(const Pair pr);
 
 Vector VertexPair_repulsion_gradient(const Pair pr);
